// src/Admin/Models/ProgramDetails.ts
export default class ProgramDetails {
  id: number = 0;
  title: string = "";
  email: string = "";
  description: string = "";
  max_reward: string = "";
  website_link: string = "";
  community_group: string = "";
  community_channel: string = "";
  clientEmail: string = "";
  program_img: string = "";
  package_id: number = -1;
  client_id: number = -1;
  client_status: number = 1;
  timeline: number = -1;
  prog_client_id: number = -1;
  leaderboard: number = -1;
  invite_ambassadors: number = -1;
  expiry_date: string = "";
}